+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.1.1
 ###### in progress
 - added minecraft 26.1 support
+- fixed server crash caused by puddle block updates running client code on the server
 
 # v1.1.0
 ###### Mar 11, 2026